Login.jsx and Register.jsx update: renamed pet image import to import RegisterPageDog from ../../assets/images/pet.png; .
